# laravel-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API – Aplikacija za izložbe fotografija",
        "description": "Swagger dokumentacija Laravel API-ja",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalno (Docker)"
        }
    ],
    "paths": {
...
        "/api/prijave": {
            "get": {
                "tags": [
                    "Prijave"
                ],
                "summary": "Lista prijava (admin)",
                "operationId": "0c614230a3d91512760eb7fef3081087",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Zabranjeno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Prijave"
                ],
                "summary": "Rezervacija mesta na izložbi (posetilac)",
                "operationId": "be50d3fcfc2ffb2443d1b0d242b94ceb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "izlozba_id"
                                ],
                                "properties": {
                                    "izlozba_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rezervisano"
                    },
                    "422": {
                        "description": "Nema mesta / validacija"
                    },
                    "403": {
                        "description": "Samo posetilac"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/prijave/{id}": {
            "delete": {
                "tags": [
                    "Prijave"
                ],
                "summary": "Brisanje prijave (admin ili vlasnik)",
                "operationId": "9b202a29bbc5850505ee74951e036ec4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obrisano"
                    },
                    "404": {
                        "description": "Nije pronađena"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/izlozbe/{id}/prijave/datum": {
            "put": {
                "tags": [
                    "Prijave"
                ],
                "summary": "Ažuriranje datuma prijave za izložbu (admin)",
                "operationId": "eca9482810b2aac9efd15d8483576763",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "datum_prijave"
                                ],
                                "properties": {
                                    "datum_prijave": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Zabranjeno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Unesi token u formatu: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}